pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "project"
version = "0.1.0"
description = "Financial Phrasebank Sentiment Analysis - Group 49"
authors = [
  { name = "Group_49", email = "arham.noman.50@gmail.com, lykostratisk@gmail.com, chinmaydongarkar@gmail.com, konstandinos.eng@gmail.com, nickiliak847@gmail.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps", "NLP", "Sentiment Analysis", "financial-phrasebank"]
readme = "README.md"


classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]

requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.6",
    "torch==2.6.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "datasets>=3.2.0",
    "transformers>=4.40.0",
    "click==8.1.7",
    "docker==7.1.0",
    "hydra-core==1.3.2",
    "dvc[gs]>=3.59.0",
    "google-cloud-storage",
    "locust==2.33.0",
    "invoke>=2.2.0",
    "dvc-gdrive>=3.0.1",
    "pathspec<0.12.0",
]

[project.scripts]
train = "project.train:app"
predict = "project.predict:main"

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
  "mypy==1.11.2",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.11"
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
omit = ["tests/*"]
